INTERLOK-3010 Move protected methods out of StandardWorkflow
Move protected methods out of StandardWorkflow into StandardWorkflowImp
which means that you can get all the behavioural characteristics of
StandardWorkflow w/o having to override the synchronised onAM (which
code analyzers don't like).
